fix compile

+++ utilities/rate_limiters/write_amp_based_rate_limiter.cc
@@ -362,7 +362,8 @@ Status WriteAmpBasedRateLimiter::Tune() {
     percent_delta_ = 100;
     critical_pace_up_.store(false, std::memory_order_relaxed);
   } else if (normal_pace_up_.load(std::memory_order_relaxed)) {
-    percent_delta_ = std::max(percent_delta_, padding / new_bytes_per_sec);
+    percent_delta_ = std::max(
+        percent_delta_, static_cast<uint32_t>(padding / new_bytes_per_sec));
     normal_pace_up_.store(false, std::memory_order_relaxed);
   }
   padding += new_bytes_per_sec * percent_delta_ / 100;
